+++ src/main/kotlin/hr/eduwalk/data/database/table/WalkScoreTable.kt
@@ -19,7 +19,7 @@ object WalkScoreTable : Table() {
             onDelete = ReferenceOption.CASCADE,
             onUpdate = ReferenceOption.CASCADE,
         )
-    val score = float(name = "score").nullable().default(defaultValue = null)
+    val score = integer(name = "score").nullable().default(defaultValue = null)
 
     override val primaryKey = PrimaryKey(userId, walkId)
 }
